Implementation of editing citizen data by import number and citizen id, when patch request comes

+++ src/main/java/dev/kolyapetrov/mypracticewithyandextask/service/ImportService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ImportService {
     Long saveImportData(Import importData);
-    Citizen editCitizen(Integer importId, Integer citizenId, Citizen eteredCitizen);
+    Citizen editCitizen(Long importId, Long citizenId, Citizen eteredCitizen);
 
     List<Citizen> getCitizensByImportId(Long importId);
 }
